Cleanup coordinator selector
