+++ stacks-signer/src/lib.rs
@@ -26,6 +26,8 @@ pub mod cli;
 pub mod client;
 /// The configuration module for the signer
 pub mod config;
+/// The coordinator selector for the signer
+pub mod coordinator;
 /// The primary runloop for the signer
 pub mod runloop;
 /// The signer module for processing events
